aggregator/requirements.txt
fastapi
uvicorn
pydantic
requests
PyJWT
python-multipart
